Extracted TerminalWidget class
